back/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.0
alembic==1.13.1
python-multipart==0.0.9
